feat: update single project route
